moved database backup\restore and added button to family-manager
